+++ ci_environment/system_info/recipes/default.rb
@@ -13,10 +13,8 @@ git '/usr/local/system_info' do
   action :sync
 end
 
-directory '/usr/local/system_info' do
-  owner node.travis_build_environment.user
-  group node.travis_build_environment.group
-  recursive true
+execute "set owner on /usr/local/system_info" do
+  command "chown -R #{node.travis_build_environment.user}:#{node.travis_build_environment.group} /usr/local/system_info"
 end
 
 directory '/usr/share/travis' do
@@ -25,10 +23,21 @@ directory '/usr/share/travis' do
   recursive true
 end
 
-bash 'Install system_info gems' do
+execute 'execute-system_info' do
   user node.travis_build_environment.user
-  cwd  '/usr/local/system_info'
-  code <<-EOF
-    bundle install --deployment
+  command <<-EOF
+    bash -l -c 'cd /usr/local/system_info
+    env FORMATS=human,json HUMAN_OUTPUT=/usr/share/travis/system_info JSON_OUTPUT=/usr/share/travis/system_info.json \
+      bundle exec ./bin/system_info #{node['system_info']['cookbooks_sha'] || 'fffffff'}'
   EOF
+  action :nothing
+end
+
+execute 'Install system_info gems' do
+  user node.travis_build_environment.user
+  command <<-EOF
+    bash -l -c 'cd /usr/local/system_info
+    bundle install --deployment'
+  EOF
+  notifies :run, 'execute[execute-system_info]'
 end