+++ worker_host/ssh/recipes/sshd.rb
@@ -1,4 +1,8 @@
+service_provider = Chef::Provider::Service::Upstart if 'ubuntu' == node['platform'] &&
+  Chef::VersionConstraint.new('>= 12.04').include?(node['platform_version'])
+
 service "ssh" do
+  provider service_provider
   supports :restart => true, :reload => true
   action :enable
 end